pyproject.toml
[tool.poetry]
name = "k8splayground"
version = "0.1.0"
description = ""
authors = ["berezzin <biarezzin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.1"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
